Switch tryjob builder to be a Tricium verifier.
We will want the tryjob builder to be able to post robot comments
back to the CL in Gerrit (in order to link to the staged site for
previews), so we need the builder to be a Tricium verifier.

This CL switches the builder over.

Bug: 1260464
Change-Id: I39e6439f0f63ab4f4029a32c73f8a871348c38b7
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/3258483
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Dirk Pranke <dpranke@google.com>