Adjust how exporting works to always use 'index.md' for files.
The older version of the export scripts would try to reduce
the number of directories we created and, so we would only
put a markdown file in a separate directory if it was the parent
of other files (e.g., /Home would be exported to
/Home/index.md, but /Home/chromecompatfaq would be exported to
/Home/chromecompatfaq.md).

This could result in multiple files being in a single directory,
which is nice as long as the files aren't referencing other
assets like images; if you add a foo.png next to the md file,
it's not clear which file references foo.png.

And, it turns out that Eleventy by default moves things around anyway
so that /Home/chromecompatfaq.md would get compiled to
/Home/chromecompatfaq/index.html.

So, this CL changes the export so that every file gets its own
directory. This is something we can change down the road, of course,
but at least this makes the initial mapping from Sites URL to file
location straightforward.

Change-Id: I4a7ecb3cbaab25236844672e44cf2911b70b9b25
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/3260348
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>