Add a full list of pages to the site.
This CL modifies //scripts/export.py to write a list of all of
the content pages it exported to //site/pages.json. This can
be useful for reviewing the site, and at some point we might
want to adapt this into a sitemap as well.

Bug: 1267642
Change-Id: I749951a1379ecad7663c39589737a52916909097
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/3270811
Auto-Submit: Dirk Pranke <dpranke@google.com>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Dirk Pranke <dpranke@google.com>