Add UI for comparing old/new pages.
This modifies the generated pages to display a "page bar" at the
top of the page that allows us to toggle between the old version
of the page (which was obtained by previously crawling the site
using crawl.py, uploading a tarball of the resulting pages, and
downloading the tarball now in the hooks), the new version,
and a "both" version that displays the two versions side-by-side.

The page bar works by fetching the full list of pages into a navlist
that you can move back and forward in, displaying the old page
in an iframe, and displaying the new page directly.

The bar is still a little buggy and if you click on links in the
pages (as opposed to using the page bar), you may navigate away
from the page (to the old site) or lose the view that you had.
These bugs are fixable and I will address them in follow-up CLs,
time permitting, but I want to get this in front of people ASAP.

Bug: 1267642
Change-Id: I06dd15100f8f3de5d4e96a5e0b9baf17b75a711d
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/3265529
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>