Add an "Edit this page" link.
This adds a link to the left nav bar that, when clicked, will
open the Git Source Editor (aka edit.chromium.org) on the
corresponding page.

The color is probably a bit garish on its own and will need
to be changed for launch, but should match the page selector
bar pretty well in the meantime.

Bug: 1267643
Change-Id: I259aecb658f86c890b4782a16d3db37b595bfc44
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/3270817
Auto-Submit: Dirk Pranke <dpranke@google.com>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Dirk Pranke <dpranke@google.com>