Clean up another 8 pages:
- /Home/chromium-security
- /chromium-os/chrome-os-systems-supporting-linux
- /developers/design-documents/create-amazing-password-forms
- /developers/how-tos/api-keys
- /developers/how-tos/get-the-code
- /developers/how-tos/run-chromium-with-flags
- /getting-involved/dev-channel

This CL tweaks the CSS styling for list items - the
Markdown generator seems to be kinda buggy and often
inserts <p> tags inside <li> tags, so we use the
`li > p:first-of-type` selector to hide that. It also
tweaks the formatting for <code> and <pre> items to
have a background color to better distinguish them
(and match the old site).

Bug: 1260513
Change-Id: Ie63fdf0d0557f649c2c0969160b6f7261500db30
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/3277232
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>