Fix default flexbox width for old/new view.
When the new version of a page is wider than 50% of the window,
in the "both pages" view, the old version was being shrunk; worst
case, if the new version was wider than 100% of the window, the
old version would appear to be missing.

This CL fixes the issue by setting both columns to a width of
50%; now, when the new page is too wide, we just get a
horizontal scroll bar.

Bug: 1270874
Change-Id: I16b3553b87412a32daf9641525abab7220b40fc6
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/3285772
Auto-Submit: Dirk Pranke <dpranke@google.com>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Dirk Pranke <dpranke@google.com>