Add support for listing the subpages of a page.
This CL adds the ability for a page to include a list of all of
its subpages (in the sense that /developers/design-documents is
a subpage of /developers) automatically, rather than maintaining
a list manually.

To use this feature, add `{% subpages collections.all %}` to
your markdown. This results in a hierarchical tree of pages
being inserted into the HTML using <li> and <details> elements.

This CL also fixes the three pages that I know of that were
using subpage listings, but there are probably still more that
need to be updated. One of the pages also just needed to be
cleaned up generally, so I did that as well while I was touching
the page.

Bug: 1260453

Change-Id: I6613f01aaa7527eeea14d2d55bbab45339a2cf89
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/3290456
Auto-Submit: Dirk Pranke <dpranke@google.com>
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Domenic Denicola <domenic@chromium.org>