Update bot config to require successful builds before submitting.
Previously the try builder was configured to run as a Tricium
verifier only, so a failure did not keep a CL from submitting.

This change modifies the config to run as both a Tricium verifier
(to still get the preview comments posted to the bug) and as
a regular CQ job (so that we are required to have a passing build
before submitting the change.).

Change-Id: I0f7b6bd283fd2e03e101bc8816e4f9c6baf7ac22
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/3291569
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>
Auto-Submit: Dirk Pranke <dpranke@google.com>