Clean up Home/chromium-security/site-isolation.
This fixes the indentation and some of the formatting for
the page.

The linked bug also talked about the styling of the table
of contents being different. I'll submit a change for that
in a separate CL since it'll affect all pages.

Bug: 1270959
Change-Id: I91cabf1051c7ff8a6914f6466b559a7175cc600f
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/3289472
Auto-Submit: Dirk Pranke <dpranke@google.com>
Reviewed-by: Gary Tong <gatong@chromium.org>
Commit-Queue: Gary Tong <gatong@chromium.org>