Change the styling of the table of contents.
This changes how the `[TOC]` tag is rendered, such that
the contents are rendered into a shaded, indented box with
a smaller font, and with the label "Contents" at the top.

This is closer to how things were rendered on the old site.

Some pages with TOCs:

- /blink
- /chrome-release-channels
- /chromiumos/developer-guide/gerrit-guide
- /Home/chromium-security/site-isolation

Bug: 1270959
Change-Id: Id2eda212207684422a32894cf70e0275d44d58ba
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/3294127
Reviewed-by: Gary Tong <gatong@chromium.org>
Commit-Queue: Gary Tong <gatong@chromium.org>
Auto-Submit: Dirk Pranke <dpranke@google.com>