Apply CSP headers when serving pages.
We want all pages to be served with CSP (Content Security Policy)
headers, to prevent XSS attacks.

In theory, at the moment, this is pointless, because the site is
completely static and there's not even any JavaScript running on
the pages (once we remove the page review bar), but this'll change
when we enable search, and probably we'll want JS for other pages
sooner-or-later, too.

Bug: 1261179
Change-Id: Ie8a349150d47b0161e51121d3a84d8d8772b15f8
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/3288316
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>