Remove page review bar.
We don't actually want to launch the site with the page review
bar displayed, so this CL removes all of that code.

We should land this only when we're ready to launch and turn
on CSP (which, in the current implementation, would break this
functionality).

Bug: 1259832
Change-Id: I38b5199c01653b2a254812ba77664bc33e367812
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/3290247
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>