Fix filenames with spaces in them.
When we exported all of the attachments that had a filename in
them, we mistakenly saved the filename with '%20' in the filename
instead of ' '.

'%20' is correct when referring to something in a URL, but a
web server will decode the %20 into a ' ' and expect that the
actual filename will have a space.

This CL fixes the issue by renaming all of the files (it also
handles two files with %2b, i.e., '+').

There are a few files that still have '%3a' (':'), which needs
to be handled as well, but ':' is an illegal character on
Windows and so we should just rename the file to something that
is safer. I'll do that in a follow-up CL.

Bug: 1273296
Change-Id: I5e8ae72c8d4d43bf391a08b91ba04a2bf0764753
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/3298770
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Dirk Pranke <dpranke@google.com>