Rename any files with '%3A' in the name.
We mistakenly exported a bunch of files using their url-encoded
filenames. For files iwth '%3A' in the name, we can't simply
use the decoded character (':') because that's illegal on
Windows. So, this CL updates the references in the web pages to
use '_' instead of '%3A', and renames the files to use '_' as well.

I thought I had fixed these when making the checkout work on
Windows, but I guess I missed some (or re-exported them :().

Bug: 1273296
Change-Id: Ibfe1afa5003510fda9543af232c69cc3f23ab5e0
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/3298771
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>