Update events pages for BlinkOn 16.
This change adds a link to the the BlinkOn 16 page to the events page,
and updates the basics of the event page itself.

This also removes a stray copy of the blinkon-16 page at the site root,
likely leftovers from the recent site conversion.

Change-Id: I3fcaa0b0d717e44535c719bc5b30eaa951c246cc
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/3536443
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Johnny Stenback <jstenback@chromium.org>