Update CHIPS origin trial documentation
In Chrome 103, we will be changing the design of the origin trial from
using the Accept-CH header to only considering the Origin-Trial header
when the response has the Set-Cookie header with a Partitioned
attribute.

The change to the OT landed in M103 in https://crrev.com/c/3679821.

Bug:1296161,1327045
Change-Id: Ifbe4ae04167ff57a573bbc2751a844feca666f74
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/3668885
Reviewed-by: Chris Fredrickson <cfredric@chromium.org>
Commit-Queue: Dylan Cutler <dylancutler@google.com>