Suggest starting spec-writing around the beginning of origin trials.
While features vary in when they should write their specs, the guidance to
finish the spec in the phase before origin trials was too early, and the
guidance lacked any instruction about when to start writing. Doing it too early
means that if the design changes, we have to rewrite an expensive spec instead
of the cheaper explainer.

I also renamed that phase to "Widen review", since "Evaluate readiness to ship"
was confusingly similar to "Prepare to ship", and the phase seems to be roughly
about getting wider reviews.

Change-Id: I15eef7476f7da8c0a04ac59a94cbf5b64c18e835
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/4162680
Commit-Queue: Jeffrey Yasskin <jyasskin@google.com>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Auto-Submit: Jeffrey Yasskin <jyasskin@google.com>