Fix (almost?) all the broken page TOCs.
The export from Google Sites was buggy and some features
did not always get exported the way we'd like.

One of those features was the way that pages' table of contents
was declared. The `[TOC]` needs to be on a line by itself
(with no other styling or anything).

This CL fixes every page except spdy/spdy-protocol/spdy-protocol-draft2/
which is quite a bit more mangled and needs to be cleaned up in a
separate CL.

Change-Id: Id4279a76f91c4983dfcfddfad7ed1e26496075e1
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/4193055
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Dirk Pranke <dpranke@google.com>