We need to propose features to working groups before shipping them.
To keep the text short, I didn't include a description of how to propose new
work to a WG, although I could add that if reviewers want. Please check that
this is clear enough that the WG doesn't need to *adopt* the feature before we
ship; we just need to propose it.

Change-Id: Id736c20881f775a69a61b3b02ac806c98e6115f0
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/4209834
Reviewed-by: Chris Wilson <cwilso@google.com>
Commit-Queue: Jeffrey Yasskin <jyasskin@google.com>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Auto-Submit: Jeffrey Yasskin <jyasskin@google.com>