chromium-os/dragonclaw: add dragonclaw v0.2 schematics
Add a copy of the Dragonclaw v2 schematics in a new dragonclaw
directory. This is a straight copy of the file already published in the
platforms/ec repository under docs/schematics/dragonclaw/, making a copy
in the website repository so we can serve it properly without asking the
user to save it locally and open that instead.

Change-Id: I14d1c1339139bdfddde23173356a7057ea840921
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/4232012
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>