Consider labelling issues as potential blockers
We've been seeing intents with many open issues, where it was hard for
the API owners to know which of these issues actually post a future
compatibility risk.
This CL adds a note to our process documentation, asking feature owners
to consider labelling their issues, to make it easier to distinguish
those that pose risk from those that do not.

Bug: 1383335
Change-Id: I3e764545c5dcdbcba63a6cb3b613d252210e6963
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/4647329
Reviewed-by: Jeffrey Yasskin <jyasskin@google.com>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Yoav Weiss <yoavweiss@chromium.org>
Reviewed-by: Rick Byers <rbyers@chromium.org>