containers-and-vms: update vmc create-extra-disk options
The removable storage option for vmc create-extra-disk is now a separate
argument at the end of the command like it is for vmc export.

Also, all host kernels have been upgraded to 4.19 or higher, so the
parts about --untrusted are no longer relevant and can be removed.

BUG=b:306398822
TEST=Run listed vmc commands

Change-Id: I117c8fcf593237c200b2a93a9b77d06f3cf8d178
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/5329238
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>