Add user guide for disabling AP RO verification
This patch adds a new guide that focuses on helping external users set
the AllowUnverifiedRo CCD flag without access to the GSC console. This
is intended to be shared with the external alternative OS enthusiast
community to help them prepare for the RO verification rollout in M122.

BUG=none
TEST=none

Change-Id: I314232796b400a09c84575b10ec7cead45c9b720
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/5371044
Commit-Queue: Julius Werner <jwerner@chromium.org>
Reviewed-by: Jett Rink <jettrink@google.com>
Reviewed-by: James Hawkins <jhawkins@chromium.org>