Make the codelab document easier to read as a Markdown document
At present, the texts has inccorect styling as a Markdown document and
the rendered website from it is hard to read.
Also, example Python codes have no indent and it's impossible to execute
with copying and pasting it.

The follwing changes improve its readability and usability.
- Remove unnecessary stylings with `**`
- Add appropriate stylings if needed
- Use code blocks for example Python codes, commands and output
- Add indents to example Python codes to make it executable
- Add blank lines to example Python codes to make it easier to read

Change-Id: Idef8535a134b891e13377187104cfe7a8e0a694e
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/5456439
Commit-Queue: James Hawkins <jhawkins@chromium.org>
Reviewed-by: James Hawkins <jhawkins@chromium.org>