Use inline codeblocks for execution commands
At present, the document also writes commands as a basic text.
It makes hard to distinguish where a command starts and ends, especially
whether a period is inside or outside a command.

To make it clear, using an inline codeblock is useful for a markdown
document.

Change-Id: Ifaa465d73fda49b484817fde5b9051ee1687a0ad
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/5462850
Reviewed-by: James Hawkins <jhawkins@chromium.org>
Commit-Queue: James Hawkins <jhawkins@chromium.org>