Sync integration checklist: Add code pointers for adding type toggles
When a new UserSelectableType is added, some manual changes must be
made to show a corresponding toggle in the UI - separately for desktop,
Android, and iOS. This CL adds code pointers for this.
While we're here, this also adds some missing code search links.

Bug: none
Change-Id: I0cf3106c752b71c19800cc934d65c2ebcd644ee2
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/5501143
Commit-Queue: Marc Treib <treib@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>